More CR fixes
