+++ src/proc/CMakeLists.txt
@@ -37,6 +37,7 @@ target_sources(${LRS_TARGET}
         "${CMAKE_CURRENT_LIST_DIR}/motion-transform.cpp"
         "${CMAKE_CURRENT_LIST_DIR}/auto-exposure-processor.cpp"
         "${CMAKE_CURRENT_LIST_DIR}/depth-decompress.cpp"
+        "${CMAKE_CURRENT_LIST_DIR}/y411-converter.cpp"
 
         "${CMAKE_CURRENT_LIST_DIR}/processing-blocks-factory.h"
         "${CMAKE_CURRENT_LIST_DIR}/align.h"
@@ -65,4 +66,5 @@ target_sources(${LRS_TARGET}
         "${CMAKE_CURRENT_LIST_DIR}/motion-transform.h"
         "${CMAKE_CURRENT_LIST_DIR}/auto-exposure-processor.h"
         "${CMAKE_CURRENT_LIST_DIR}/depth-decompress.h"
+        "${CMAKE_CURRENT_LIST_DIR}/y411-converter.h"
 )
