+++ src/proc/color-formats-converter.h
@@ -61,30 +61,4 @@ namespace librealsense
             color_converter(name, RS2_FORMAT_RGB8, RS2_STREAM_INFRARED) {};
         void process_function(byte * const dest[], const byte * source, int width, int height, int actual_size, int input_size) override;
     };
-
-    class y411_converter : public functional_processing_block
-    {
-    public:
-        y411_converter(const rs2_format & target_format)
-            : functional_processing_block("Y411 Transform", target_format) {};
-
-    protected:
-        y411_converter(const char * name, const rs2_format & target_format)
-            : functional_processing_block(
-                name, target_format, RS2_STREAM_COLOR, RS2_EXTENSION_VIDEO_FRAME)
-        {
-            _stream_filter.format = _target_format;
-            _stream_filter.stream = _target_stream;
-        }
-
-        void process_function(byte * const dest[],
-            const byte * source,
-            int width,
-            int height,
-            int actual_size,
-            int input_size) override;
-    };
-    void unpack_y411(byte * const dest[], const byte * s, int w, int h, int actual_size);
-    void unpack_y411_sse(byte * const dest, const byte * s, int w, int h, int actual_size);
-    void unpack_y411_native(byte * const dest, const byte * s, int w, int h, int actual_size);
 }
